backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.2.1",
    "chromadb>=0.5.0",
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "langchain-text-splitters>=0.2.0",
    "ollama>=0.2.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
    "sqlmodel>=0.0.19",
    "uvicorn[standard]>=0.29.0",
    "weasyprint>=62.0",
]
